test: Add HTTP visibility annotation to echo-other-node
Add HTTP visibility annotation to the echo-other-node pod. With a new
enough Cilium agent those annotations will show also when
echo-other-node has an ingress policy. Test behavior is not changed,
but we could change to verify the additional visibility when Cilium
agent with the visibility annotation is released.

To see the annotations use e.g.:

$ cilium connectivity test --test echo --print-flows --debug

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>