test: Track last match in 'LastMatch'
Track last match of any matches in 'LastMatch' so that we have the
true range of matches from 'FirstMatch' to 'LastMatch' also when flow
aggregation skips the match for 'Last' requirement. Without this the
merging and printing can skip significant flows as the full span of
significant flows was not tracked properly.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>