test: Add Path and L7 drop support
Expose URL path to policy expectation logic. This helps returning
expected test outcomes when an HTTP policy filters by path.

Add filter for L7 drops.

Add new action to tests "pod-to-world" scenario where the target is
"google.com/index.html". Without policies this should work, but in
"to-fqdns" and "egress-l7" test cases this is not allowed by the
egress policy.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>