ginkgo: Update to 1.16.4
The following controller-tools update pulls in a new ginkgo version
with a deferred Describe() change that is incompatible with the Cilium
Ginkgo suite handling of scopes. Separate this change and fix to this
commit to make it clear what is going on.

To properly deal with deferred Describe() handling of Ginkgo, we also
need to defer the scope handling by moving it to the function body
called by Ginkgo.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>