service: Add L7LBInfo to track L7 LB service registrations
Add support for ingress Cilium Envoy Resources CRDs by not installing
datapath L7LB redirection when 'Ingress' is set to true. This allows
the regular eBPF LB to be used for pod-to-pod traffic while L7 LB is
used for ingress. Service backends are still synced to Envoy, so that
ingress can load balance to them.

Keep track of the CiliumEnvoyConfig that requested a service to be
registered for L7 load balancing. Only one CiliumEnvoyConfig may do
this in egress direction, as otherwise the choice between Envoy
listeners would be ambiguous. Without this tracking two CECs could
register the same service for L7 LB redirection and removal of one of
them would remove the datapath redirection.

Any number of CECs in 'ingress' direction may register a given service
for L7 load balancing. In the ingress case this does not involve
redirection of traffic egressing from pods, so there is no
ambiguity. Instead, only the service backends are configured as Envoy
resources so that the ingress load balancer may route traffic to them.

Note: 'Ingress' CEC configurations are still experimental.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>