hubble port-forward: Remove dependency on `cilium-cli-helm-values` secret
If cilium is installed via helm, the `cilium-cli-helm-values` secret is
missing. This causes the `cilium hubble port-forward` command to fail,
since the factory method used to create a `K8sHubble` instance was
loading it.

Since the port-forward command doesn't use any of the values from the
secret, it has now been refactored to only use the already filled
`Parameters` instead of `K8sHubble`.

fixes #904

Signed-off-by: Patrick Pichler <git@patrickpichler.dev>