uninstall: set "cni.uninstall" to true
There is a Helm variable, `cni.uninstall`, that tells the agent to clean
up its CNI configuration file and plugin binaries. This behavior was
disabled by default in PR https://github.com/cilium/cilium/pull/24009.
However, when we *know* we're uninstalling Cilium, we should re-enable
the old behavior.

So, manually touch the key file in the configmap to enable uninstall.
Then, the pre-stop hook will be activated on shutdown and clean up as
usual.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>