connectivity/check: Add CAP_NET_ADMIN for host-netns ds
This capability is needed to have the ability to add a static route.
The route will route traffic from the node without Cilium, to the other
nodes' pod CIDR so that the pods are reachable from the node without
Cilium.

This is needed for the tests which assert ingress from-cidr
connectivity to pods.

Signed-off-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Martynas Pumputis <m@lambda.lt>