renovate: add renovate/stop-updating label on renovate's PRs
When renovate opens a PR, it will keep updating the dependencies as long
as the PR is open. This makes it difficult have a green CI run since
some dependencies are being updated frequently and we aren't able to
merge the PR. Thus, add this label to prevent renovate from updating PRs
after opening them.

Co-authored-by: André Martins <andre@cilium.io>
Signed-off-by: Tobias Klauser <tobias@cilium.io>